Moved the new decoder to its own class
